+++ data/order.sql
@@ -8,7 +8,7 @@ CREATE TABLE `order`(
   `type` varchar(8) NOT NULL,
   `signal_type` varchar(16) NOT NULL,
   `date` date NOT NULL,
-  `price` decimal(20,10) NOT NULL,
+  `price` double NOT NULL,
   `quantity` int NOT NULL,
   `result_type` varchar(16) NOT NULL,
   `result_price` decimal(20,10) NOT NULL,
